tools/live_plotter/requirements.txt
matplotlib
PyQt5
paho-mqtt